Added a question js file This allowed me to try and clean up the index js Now all the long blocks of question code are in a separate file.
